Admin felület működik. Fordítani kell a sorrendet az üzeneteken.

+++ src/main/java/com/beadando/javabeadando/MessageRepository.java
@@ -2,5 +2,5 @@ package com.beadando.javabeadando;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MessageRepository extends JpaRepository<Message, Long> {
+public interface MessageRepository extends JpaRepository<Message, Integer> {
 }
